+++ sandbox/vue-library/.storybook/main.ts
@@ -1,0 +1,12 @@
+import type { StorybookConfig } from '@storybook/vue3-vite';
+
+const config: StorybookConfig = {
+  stories: ['../src/stories/**/*.stories.@(js|jsx|mjs|ts|tsx|mdx)'],
+  addons: ['@storybook/addon-docs'],
+  framework: {
+    name: '@storybook/vue3-vite',
+    options: {},
+  },
+};
+
+export default config;