+++ sandbox/vue-library/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @rei/c1-vue-library Changelog
 
+## 0.1.2-alpha.0 - 2026-01-28
+
+- Added Vue Storybook config and fixture stories for local previewing.
+
 ## 0.1.1-alpha.0 - 2026-01-28
 
 - Converted the Vue fixtures app into a reusable Vue library package.